requirements.txt
fastapi
uvicorn
python-telegram-bot==20.8
openai==0.28.0
Pillow
python-dotenv